Result class has been renamed to TaskModel. Significant changes has also been made to the class.
The types of task, status, priority and the new field checked implement the interface ObservableValue<>. The reason behind this is because the interface implements listener, also the PropertyValueFactory class being passed to the setCellValueFactory methods are not suggested and should be avoided. Since the checked column uses the BooleanProperty, I've also changed the variables of type String to StringProperty.